hsv and colorizer tests save to mp4
To help with diagnostics and performance tweaking, realsense-colorizer and realsense-hsv-viewer now save to mp4.

I ran into issues earlier with invalid mp4s getting generated, the solve to get this to work was to properly terminate / end the stream. 

To get valid mp4 use the "ESC" key to exit the app instead of terminate/kill.