Fix incorrect hsv to rgb conversion
This fixes the issue where depth data seems to "wrap around" and the full 4 meters of depth is not available.  The furthers points showing up closest to the camera. 

The problem was caused by incorrect assumptions as to what the cv2.COLOR_HSV2BGR conversion takes as input. It expects a range of 0-180 and it was being supplied a range of 0-360.